+++ dino/src/game/score/mod.rs
@@ -11,6 +11,7 @@ pub struct ScorePlugin;
 impl Plugin for ScorePlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<Score>()
-            .add_systems(OnEnter(AppState::InGame), insert_score);
+            .add_systems(OnEnter(AppState::InGame), insert_score)
+            .add_systems(OnExit(AppState::GameOver), reset_score);
     }
 }
