+++ dino/src/game/score/systems.rs
@@ -4,3 +4,7 @@ use bevy::prelude::*;
 pub fn insert_score(mut commands: Commands) {
     commands.insert_resource(Score::default());
 }
+
+pub fn reset_score(mut score: ResMut<Score>) {
+    score.value = 0
+}
